refactor(scan-nh): improve clarity of helper function
I paired on this with @jonahkagan. He initially struggled to understand why we had `fold_with` that appeared to be doing a sort of reduce, but then we did a `reduce_with` after it. I believe this is happening because `fold_with` is a parallel iterator that operates on groups and that the multiple results are per group. We looked into how this could be improved and found:
1. using a parallel iterator with the number of items we had does not improve performance
2. we could better express the algorithm by getting all pairs first, then mapping each pair to a list, then picking the longest list

The new version is approximately the same speed but much clearer.