chore(scan-vx): remove deprecated `Interpreter#interpretTemplate`
This no longer works properly because of the change to require a contest offset when interpreting templates, so it's best to just remove it. There wasn't a very easy way to fix the interpret CLI so I just removed it since it's mostly used for testing/debugging (by me) and we're likely scrapping the VX ballots in their current form anyway.