feat(cental-scan): use NH-next interpreter
This moves the adapter layer from `apps/scan` to `ballot-interpreter-nh-next` itself so `apps/central-scan` can use it, then replaces the `interpret` call from `ballot-interpreter-nh` with the new compatible one from `ballot-interpreter-nh-next`.

This doesn't remove the dependency on `ballot-interpreter-nh` because it wasn't clear that it was safe to do so since it's still used in `store.ts`, though the code that uses it is slated to be removed and may itself be unused.