feat(grout): Serialize Dates (#3239)
* feat(grout): Serialize Dates

* test(admin): Fix date format in tests

Because Grout previously didn't support serializing Dates, there was one case where a Date was being implicitly serialized to a string. Luckily, the only consumer of this endpoint on the frontend was just wrapping the date string in a `moment` so there was no functional issue, but the type was incorrect. I simply updated the tests to reflect the now correct serialized Date type (which should also work fine with `moment`).