build(deps): ensure specific packages have one version (#3268)
* build(deps): ensure specific packages have one version

Also includes a change to `validate-monorepo` that ensures we don't regress. This should fix issues with VS Code + ESLint loading incompatible versions of ESLint which caused it to stop working in the editor.

* build(deps): remove `preact`

We weren't actually using it.

* test: upgrade `fast-check`

This isn't the latest version, but it's the version `central-scan` had before standardizing on the more common 2.18.0. Unfortunately, that version had an issue with the `json_serialization` tests so this bumps it up a little bit.