perf(scan-nh): make returning normalized images fast
Previously we were including the normalized images as part of the interpretation result that was serialized as JSON on the Rust side and deserialized on the JS side. This took much longer than the interpretation, adding about 1s on my M1 MAX MBP, and it also generated a string that was several hundred megabytes. Even after we stopped pretty-printing the JSON it was still over 100MB. Instead, we now create `JsBuffer` objects that hold the raw image data and return that directly without modifying it. There's more we can do to improve this, but it requires modifying the whole scanned image handling pipeline.