feat(nh-next): Return normalized images and computed ballot layouts (#3267)
* doc(nh-next): Tweak README install instructions

* feat(nh-next): Return normalized image with results

* refactor(nh-next-adapter): Use side to determine page number

This was originally a commit to fix this issue, but then it got fixed out from under me by @eventualbuddha. I left in my version of the fix as a refactor because I like how it reads, but I'm not attached to it.

* feat(nh-next): Compute and return ballot layout

* fix(scan): Export layout from interpretation result for VXF CVRs

* use computed layouts for CDF export when available. have not added testing

* style: apply some suggestions from `clippy`

* perf(scan-nh): make returning normalized images fast

Previously we were including the normalized images as part of the interpretation result that was serialized as JSON on the Rust side and deserialized on the JS side. This took much longer than the interpretation, adding about 1s on my M1 MAX MBP, and it also generated a string that was several hundred megabytes. Even after we stopped pretty-printing the JSON it was still over 100MB. Instead, we now create `JsBuffer` objects that hold the raw image data and return that directly without modifying it. There's more we can do to improve this, but it requires modifying the whole scanned image handling pipeline.

* PR feedback

* Layout and normalized image tests

In order to make the tests deterministic, I refactored the layout
computation to use a stable ordering of contests based on gridLayouts.  Probably good to maintain that order anyway.

---------

Co-authored-by: adghayes@gmail.com <adghayes@gmail.com>
Co-authored-by: Brian Donovan <1938+eventualbuddha@users.noreply.github.com>