fix(scan): use correct layout for grid-based elections
For some reason we still generate VX-style ballots when building ballot packages for grid-based elections, which means that a layout is generated too. That layout is unlikely to put contests on the same pages as the real layout, which can result in contests appearing on the wrong pages. This affects write-in adjudication because a write-in may be ignored if the two layouts disagree about which page it is on. This commit fixes the issue by always using the generated layouts for grid-based elections, ignoring the vestigial VX-style ballots generated for such elections.