feat(scan-nh): allow interpreting `ImageData`
Previously we required passing paths to image files to load. Now we can pass `ImageData` objects directly. This has the potential to be faster, though in my basic testing it was only faster some of the time. I'm hoping that eliminating the encode+write of the images does make a difference, though.