refactor(auth): use `tmp` directly (#3300)
* refactor(auth): use `tmp` directly

Rather than implementing our own `tmp`-like functionality, let's just use the `tmp` package. Advantages:
- we don't have to hardcode a particular place to put files
- we don't have to implement logic to avoid writing to an existing file
- we can use the existing `tmp` cleanup facility

All of the above simplifies the tests somewhat as well.

* refactor(auth): work around cypress test issue

Cypress is importing `DEV_JURISDICTION` by referencing a `.ts` file directly. This works because webpack is transpiling anyway, but the `certs.ts` file it was importing now imports the `tmp` package indirectly. This fails since the polyfilled `fs` package webpack provides doesn't have a `constants` exports, so we get "`O_CREAT` is not a property of undefined" or some such error. This commit works around the issue by putting the constants in a `constants.ts` that the test can reference.