feat(scan-nh): allow interpreting `ImageData` (#3317)
* refactor(scan): interpret images, not paths

Internally we now accept `GrayImage` instead of `Path` in `interpret_ballot_card`. The JS interface still accepts string paths, but that'll change to also accept `ImageData` objects in the next commit.

* feat(scan-nh): allow interpreting `ImageData`

Previously we required passing paths to image files to load. Now we can pass `ImageData` objects directly. This has the potential to be faster, though in my basic testing it was only faster some of the time. I'm hoping that eliminating the encode+write of the images does make a difference, though.