fix(scan): Prevent double sounds on accept
It's possible for the scanner status from the state machine to change
even when the state doesn't change - in this case, the `interpretation`
field was being cleared before leaving the `accepted` state. This caused
the success sound to play twice. To fix, we make sure we only play
sounds when the actual state changes.