feat(scan): remove uses of `UninterpretableHmpb`
We no longer use this adjudication reason, so we should no longer build election definitions that use it. However, this commit doesn't remove it completely because doing so has a lot of negative effects: all elections that use it become invalid; all fixtures need to be updated, including ballot packages; many tests need to be updated. We should do that work eventually, but for now I wanted put a line in the sand to stop using it.