feat(basics): add early-return for `Result` functions (#3367)
* feat(basics): add early-return for `Result` functions

Rust has the `?` operator to return early from functions if the `Result` operand is `Err`. This is convenient, and the lack of it with our `Result` type means a lot of extra error checks with early returns. This is annoying boilerplate, but it also makes complete code coverage difficult to achieve since you need to write tests that cover what is essentially just propagating an error.

A note about the design of this API: it uses JavaScript exceptions. This is not ideal but is unavoidable since it's the only way to return early from a function without `return`. Additionally, you might ask why `resultBlock` has a function to be passed to `or` rather than simply having `or` (or some other `Result` method) throw the exception. The reason for that is to ensure the errors propagated have the right type. Otherwise a thrown `Err` might not be compatible with the type expected by the containing function.

* refactor: replace `mapResult` and some early returns

This uses the new `resultBlock` and `asyncResultBlock` functions to make the code follow a more readable style of error propagation.

* refactor(basics): use more generic `okOrElse`

Turns `Result::or` into the more generic `Result::okOrElse`. Also changes the callback name from `ret` to `fail`.