ci(utils): use `v8` coverage provider
Switches from `istanbul` to `v8`/`c8` coverage provider to avoid running out of memory. There's an issue with `istanbul`'s coverage when run with Jest that we haven't quite nailed down but seems to be related to sourcemap generation. Rather than digging into that, this PR continues down the road `admin/backend` started down of using `v8` instead. The plus side is that it's much faster and doesn't run out of memory. The downside is that it's not as smart as istanbul. It operates on the compiled JS, which may sometimes mean that a bit of code isn't covered that doesn't actually have a semantic equivalent in the original TypeScript code. It also has dumber "ignore" directives, where you can specify a start/stop and a number of lines, but it doesn't operate semantically by statement or `if`/`else` branch.