Add apps/design/shared package skeleton
There's some code that will need to be shared between the frontend and
backend (e.g. the logic for laying out a ballot). In the past, we've
created packages under /libs for logic like this. Since this code is
specific to apps/design, I thought I'd experiment with keeping it under
the app directory structure.