Ballot auto-layout (first draft)
Move some of the existing document rendering primitives from apps/design/frontend to apps/design/shared.

Tweaks the primitives for rendering an actual ballot:
- Adds TextBox
- Adds Image
- Collapses Box and Rectangle into just Rectangle
- Removes Ellipse

Then uses these primitives to create a function that lays out a ballot for a given ballot style. The function returns a document (for rendering) and gridLayout info (for exporting a ballot definition that will configure a scanner).

In this first draft, I cut a lot of corners and hardcoded a lot of things. The process of working through this first iteration helped me figure out what a better approach to auto-layout would look like. I haven't pursued it yet, since it's not high priority at the moment. But suffice to say I don't expect this particular approach to survive very long, so I wouldn't spend too much time grokking it.

The high-level idea that I do think will stick around is to compose functional components (a la React). I think this worked out fairly well.