admin: Fix buffer overflow when zipping large ballot packages (#3606)
When trying to zip a large election definition file into a ballot
package, the zip-stream library was hanging. My hypothesis is that this
was due to the stream trying to buffer all the data, since we did not
pipe it to any sink before adding the file contents to the stream.

To fix this, we pipe to an output file write stream before adding any
files to the stream.

I also added some error handling just in case.