Encode ballot style and precinct in metadata timing marks by index
We currently use the `cardNumber` field in the metadata to identify
ballot styles of the format "card-number-{n}". This is how we support
interpretation for NH Accuvote ballots. We only support a single
precinct in this case.

For VxDesign ballots, we need to encode both the ballot style and the
precinct. Furthermore, we want to support ballot styles with
arbitrary formats (since VxDesign accepts arbitrary ballot styles in its
input election definition). So we add logic to treat the `cardNumber`
field as an index into the ballot styles list, as well as encoding the
`precinctOrBatchNumber` field to be an index into the precincts list.

Note that this is just a temp solution for m17 to maintain backwards
compatibility. The long-term solution will be to encode this data in a
QR code on the ballot.