style(jsx): remove unnecessary `React` from scope (#3658)
* style: do not require `React` to be in scope

Per the docs on `react/react-in-jsx-scope` [1], the rule is only needed when not using the `react-jsx` runtime, which we are using. So we disable it by extending the `jsx-runtime` config as described in the note at the bottom.

[1]: https://github.com/jsx-eslint/eslint-plugin-react/blob/master/docs/rules/react-in-jsx-scope.md

* style(jsx): remove unnecessary `React` from scope

Now that we no longer enable the `react/react-in-jsx-scope` rule, we can get rid of these imports. Much of this was done with search and replace.