Backend: Creating/editing elections
- Adds backend API endpoints for creating/editing/deleting elections.
  Includes support for managing multiple elections.

- Overrides the stored election.precincts with a custom precinct type
  that supports split precincts. This seemed like the quickest/least
invasive way to prototype support for split precincts (rather than
changing the Election type). The custom precinct type stores which
districts each precinct/split belongs to.

- Based on this data, we can auto-generate ballot styles. In the VXF
  format, ballot styles encoded how different precincts were split based
on district boundaries. In this data model, the user enters precinct
splits and which districts they correspond to. We then use this info to
create the corresponding ballot styles.

Sorry, I know it's a lot for one commit!