Restructure frontend app routing
The previous iteration of VxDesign had just a few routes and pretty much everything was in app.tsx. This commit sets up a more complex hierarchy of pages. Check routes.ts for a skeleton of the routes that should make the structure clear.

This commit mostly removes code from app.tsx. Upcoming commits will add new files for different sections of the hierarchy.