ElectionInfoScreen
The first step of defining the election - name, date, etc.

In general, the pattern I'm trying to follow is having all data first be
displayed in a "saved" state. I think this makes the app feel more
trustworthy (vs always showing an editable state). That way, you know
that your inputted data is securely saved.

This commit establishes what that pattern looks like for a basic form.
The form starts in the saved state, then the user clicks "Edit" to turn
it into an editable form.

I did make an exception for when the form is empty and started that in
an editable state, since it was confusing to click "Create Election" and
then be shown a locked form.