GeographyScreen
For editing districts and precincts (incl. splits).

This establishes the second pattern, used for collections of data. The list of saved items is shown in a table. The user can click "Add" or "Edit" to go to a form for a single item.

I chose to use a separate page with breadcrumbs at the top for navigation rather than a modal because these forms can get pretty complex. I wanted full screen real estate.