feat(scan-nh): optionally count unmarked write-ins (#3704)
Refs WIA: ability to adjudicate name-only marks (no oval) #2458

If `markThresholds.writeInTextArea` is set, the NH interpreter will now compute a score for the dark pixels in each write-in area. If the score crosses the threshold, the fill score for that write-in will be set to 100% (1). This is an interim solution for detecting unmarked write-ins that we may decide to keep as one of multiple different configurations, e.g. alongside including such ballots in the write-in adjudication flow.

To enable this feature, simply populate `markThresholds.writeInTextArea` within an NH election definition with a value. I recommend 5-7% based on my testing.

The interpreter does not automatically score write-ins and instead makes that an option that is currently enabled when the threshold is set. I structured it this way primarily because the interpreter does not use the knowledge of the mark thresholds and I'd like to keep it that way. It also doesn't modify the score as described earlier, but returns the write-in scores separately. Setting the score to 100% happens at a higher level of the interpretation stack.