Encode ballot metadata in QR code
Restore the libs/ballot-encoder logic for encoding HMPB metadata. Then,
using the qrcodegen library, generate the QR code data. Finally, render
this data using our normal ballot rendering methods.

Note that the qrcodegen library is the same library that underlies
qrcode.react - the library we use for generating BMD QR codes. It is not
released on npm - you have to copy the source file in yourself.