public-hoist all eslint packages (#3781)
I'd like to make our dependencies easier to understand. Currently, all of our package.json files reference a bunch of eslint configs and plugins that they don't explicitly use in their eslint configs. Why? Because they are referenced in eslint-plugin-vx and eslint is not able to resolve the packages if they're not in a flat node module structure (like npm). To replicate this with pnpm, we "publicly hoist" all eslint packages. This is the default config that pnpm now ships with. All publicly hoisted packages exist in a flat node modules directory at the root of the repo.

My goal is two-fold:
- make the package.json more informative by only including explicitly referenced packages
- avoid drifting versions of eslint. there were places in the repo where we had upgraded plugin dependencies of eslint-plugin-vx but hadn't upgraded those plugins in other packages, meaning linting was not getting bug fixes / improvements in those packages as intended