Hash election after computing gridLayouts
In order to have the correct election hash encoded in the ballot
metadata QR code, we have to first lay out all of the ballots to compute
gridLayouts.

To support this change, we also undo passing an ElectionDefinition
around in VxDesign, since we only need the Election data until we're
finally ready to layout the ballots and compute the hash.