Cache all HMPB fixtures
Generating the ballot PDFs was the slowest part of the interpretation
tests. So we follow the pattern of all bubble ballots and pre-generate
these fixtures, caching them in libs/hmpb/render-backend.

In doing so, I opted to tweak the fixtures to have a common pattern of
what data they export.