Update converter-nh-accuvote, hardcoding sheetNumber to 1
We currently implicitly only supported single sheets, so maintaining
that behavior for now unless we have a reason we want to support
converting multi-sheet NH ballots. I peeked in the XML fixtures and
didn't see any field that would tell you the sheet number or page
number, so it would take more work to fiure out how to do this anyway.