Remove PrecinctScannerInterpreter abstraction
This was created when we needed to cache interpreter config data that was expensive
to load from the database (physical ballot template PDFs). Now, it's
cheap to load all of the config for the interpreter on every scan. And
in fact the code was already doing this, but the old abstraction hadn't
been cleaned up.

I reduced the interpreter object to a simple interpret function, which
is still dependency-injected for easy mocking in tests.