Remove BallotLayout.targetMarkPosition and BallotLayout.layoutDensity
VxSuite doesn't use these fields for interpretation, but VxDesign uses
them for ballot layout. To ensure that the election exported is as
minimal as possible, I'm removing these fields. (Also partially
motivated by the fact that they don't exist in CDF, so it's an easy
place to accidentally make mistakes when converting.)