VxAdmin: Store system settings with election
System settings are election-specific, since they are imported at the same time (in the election package). Since the backend of VxAdmin supports storing multiple elections, it makes sense to store the system settings with each election.

In making this change, I also thought it made sense for the API to
configure the election to reflect this change as well.