perf(auth): read file data as needed
Rather than reading all the file data for a given directory upfront, read it at the point where we actually need it to compute the hash. Additionally, we don't need to read all the file data for a given file all at once. It can be streamed while we compute the hash incrementally.