Clean up election fixtures (#3912)
* libs/fixtures: Remove deleted resources when building resources

* Clean up outdated resource files using updated build_resources.sh script

* Remove primary contests/ballot styles from electionSample.json

* Regenerate HMPB fixtures

Due to election hash change

* Update mark/mark-scan snapshots

* VxDesign: import elections with precincts with no districts assigned

electionSample now has a stray precinct, which should be allowed

* Remove electionPrimarySample fixture and update corresponding tests

It was barely a primary - other fixtures represent the primary use case
better. And it was barely used.

* Remove unused electionMinimalExhaustiveSampleRightSideTargets fixture

* Remove electionSample2 fixture

Wasn't being used for anything meaningful that another fixture couldn't
do

* Remove electionSampleCDF fixture

I was just creating it using the conversion code anyway, so no reason to
cache it.

* Remove sampleAdminInitialSetupPackage fixture

Keep the systemSettings.json file, but move it to the top level. Get rid
of the election, which was just a copy of another election fixture and
wasn't being used.