fixes an issue that wouldn't occur in prod, but occurs in dev
we don't actually ever delete anything from the VxAdmin DB. when elections are deleted, they are just marked with a deleted flag. this is something we should probably fix before we forget about it and cause a catastrophe of laptops filling up their storage

anyway, as such, if you delete an election and reload a CVR file, the batch IDs are the same. and the insert operation there is "insert or replace", so the batches end up still being pointed to the old election marked for deletion

by changing to a composite key specifying the election, the "insert or replace" operation creates a new record every... new election