build: upgrade to TypeScript 5 (#3960)
* build: upgrade to TypeScript 5

Also updates related packages, especially around linting. Both the TS update and the ESLint updates required some changes.

* fix: remove `codemods`

It had a number of incompatibilities with TypeScript 5 and it's easier to just remove it since we don't actively use it day to day.

* fix: various changes for newer `@typescript-eslint/*`

- `parser` requires absolute path to parser module
- `typeParameters` is now `typeArguments`
- `RuleTester` was extracted to its own package

* ci(eslint-plugin-vx): ignore added early return

This was added due to a change in the type of `elements` to include `null` items, but I don't know when this would happen. Maybe an elided array entry?

* test(image-utils): add missing test for `ensureImageData` branch

* ci: upgrade `cypress`

I'm running into issues with TS config `moduleResolution` in the integration tests in CI, so it may be that newer versions have better TS 5 support.

* test(mark): switch from Cypress to Playwright

It's faster and better supported. In particular, it doesn't have the problems Cypress has with TypeScript v5 (vis a vis webpack, I think).

* ci(playwright): install `chromium` before testing

* ci(playwright): fix `install` command invocation

* ci(playwright): install deps & cache tools

* test(admin): replace Cypress with Playwright

* test(mark-scan): switch from Cypress to Playwright

* test(integration): switch everything else from Cypress to Playwright

* ci: restore `central-scan` integration test job