fix(hmpb): improve QR code detection (#4092)
* fix(hmpb): improve QR code detection

Changes our QR code detection code in Rust to be closer to our code in TypeScript. We now use multiple QR code readers, specifically `rqrr` (slightly faster) with fallback to `zbar` (better detection, but not complete overlap with `rqrr`). Additionally, we now only look at the parts of the image that we know might contain a QR code: the top-right and bottom-left corners. This speeds up the detection a lot, dropping from 60-150ms on my M1 MAX to 2-12ms.

The crop factor of 20% width × 20% height was not chosen terribly scientifically. I just deemed it a reasonable crop with enough error margin based on looking at a few scans.

* ci: update docker images to include `libzbar-dev`

* perf(scan): have `zbar` look for only QR codes

This doesn't seem to make much difference on the data set I tested with, but it still seems like a worthwhile thing to do.

* chore: make clippy happier

* refactor: use `thiserror` for better errors