Add desktop palette and color theme
Creates a new extended set of color theme roles to support more complex
themes. Adapts the existing color themes to this new type (while leaving in the
legacy color theme fields for backwards compatibility during the
transition).

Also adds the concept of a color palette, which contains a set of colors
to assign to roles in a theme.