Button cleanup (#4144)
* Remove deprecated Button styling props

They don't actually influence the button styling anymore, so there's no point in keeping
them around. Also remove deprecated ButtonInterface.

* Remove DecoyButton and Button.component prop

These were in use on the Review screen to make the entire contest card
pressable (with a fake button inside of it). I found a different
solution: using the regular onClick handler for the contest component
and adding a similar handler to the button inside of it. I also had to
add a tab index to the inner button to replicate the screen reader
behavior of the original approach, which is to read the entire contest
card as one unit and skip the inner button when tabbing through
elements.

* Use right icon for start voting button