Voter Privacy Warning / Tally Report Viewer Refactor (#4150)
add warning types and warning text generation

previously there were a lot of paths to generate data. you could load the preview first, and then print or export it. or you could directly print or export, which would have a side effect of generating the preview. or you could export the .csv, which would not generate any preview.

altogether it was too complicated and footgun-y. now, you have to click “Generate Report”, which shows the preview and enables the other options. only drawback is that when exporting .csv, you end up running the query twice. but what this means we need is backend caching (if anyting), not a hacky frontend