Update SegmentedButton (#4159)
This required also tweaking Button, which previously rendered all disabled buttons with the same background in the touchscreen theme. This doesn't work well for segmented buttons, which need to show which button is selected even when disabled. So I nixed that and removed the background color changes for disabled buttons in touchscreen themes.