Desktop UI components (#4160)
* Remove fractional percentages from HSL colors

The `polished` library we use in various places can't parse these, and they aren't necessary. They were an artifact of the color palette generation tool I used initially.

* Add spacing guidelines

* Tweak Button styles to follow spacing conventions

Also fix a bug introduced by the `rightIcon` prop that caused icons on
the right to wrap below the button label, which was inconsistent with
how left icons wrapped above the button label.

* Update ButtonBar

* Update Card

* Update ContestChoiceButton

* Update DisplaySettings

* Update ElectionInfoBar

* Update SVG images

And move images to grouped story

* Remove flex item settings from Button text

Otherwise, when the button gets stretched (e.g. to fill a button bar in a modal), the icon gets separated from the text because the text
container grows.

* Update Modal

I removed the border because it was lower contrast with the background overlay (across all themes) which made a less clear delineation of the edge of the modal.

* Update Seal

* Update GlobalStyles

* Remove unnecessary nesting from UnlockMachineScreen story

* Update TabbedSection

For now, not worrying about making this work well on desktop because I don't think we currently use it there. Can revisit later if we want.

* Update TouchTextInput

* Update WithScrollButtons

* Update a few tests

* Add border radius to size theme

* Update comment in Icons