fix(ballot-interpreter): improve QR code detection (#4203)
Changes the search area bounds to better handle skew. In comparing the search area with the actual QR code location, I found that the area was sometimes not wide enough and also often taller than needed. This also changes the height of the search area to be based on the ballot image width, since the QR code is anchored to the bottom of the image and searching more in a taller ballot will not help.

This has a negligible impact on performance. For a letter-sized ballot image (1700x2200 pixels), we previously were searching 149,600 pixels and we're now searching 180,625 pixels. This did not seem to have a noticeable impact on runtime.