Update System Administrator screen
Instead of using the shared component SystemAdministratorScreenContents,
build out our own screen. I chose to do this because the shared
component is just a centered grid of buttons, which is not very
user-friendly and breaks from our overall pattern of left-alignment in
desktop apps. Open to feedback on this choice.