Update unconfigured election screen
This screen is shown when logged in but without a USB drive. In order to
get our normal controls (e.g. lock machine), we use NavigationScreen
with no nav items. It's a little weird because the contents of the
screen are still centered, but I think it's good enough for now.