Change medium contrast warning color
Due to the 10:1 contrast ratio, it's impossible to find a warning color that actually looks orange/yellow, it just comes out brown. Due to this constraint, we had originally decided to just use the main text color for warning accents. I accidentally changed that somewhere along the way. This PR resets it.