perf(scan): stop copying scanned images in memory
I added the `trimBlackFromTopAndBottomOfImage` function when I first added the Custom state machine because the images we got back from the scanner had somewhat tall black header and footer portions that interfered with the interpreter as it was then. Maybe. It's not entirely clear because _this function doesn't seem to do what it says it does_. It's looking for rows at the top and bottom of the image that it can trim off, but in practice it always seems to trim zero rows on both ends, effectively just copying the images. We can tell it's not trimming because the `imageHeight` calculation of the trimmed image is wrong: it's using `image.imageBuffer.length / offsetPerRow`, but the numerator should be the _new_ length, not the _old_ length. If it actually were cropping any rows then the width × height of the trimmed image would not equal the number of byes in the buffer and `createImageData` would reject it. (In trying to make the function _actually_ do cropping, I had exactly this happen.)

I believe maybe the old TS interpreter didn't handle the header/footer but the Rust one does. I'm not sure, and I don't think it's worth digging too deeply into.