fix(utils): ensure `Mutex` internal async queue is cleared
@arsalansufi discovered that the private `asyncQueue` property of `Mutex` only ever grew and was never drained. This isn't great, and represents a memory leak. The only question is whether it's a fast or slow memory leak, which @arsalansufi is investigating.

While I was here I added some more testing, especially testing that covers the cases we care about: either with a value or without a value. For the case without a value I made it so you can do `new Mutex()` and just get one whose value is `void`. This avoids needing to give it a dummy value like `0`.