fix(utils): further fixes to async locking
We were re-using the `deferred`, which meant that if it had to be awaited more than once it would not actually wait for `unlock`, only yield to the run loop. This meant that the `for (;;)` loop would spin quite quickly and add a lot of entries to `this.asyncQueue`, eventually causing a crash due to allocation failure.

Paired with @arsalansufi on this.