Check for mismatched precinct, ballot style, or page numbers
When interpreting a VX HMPB, we should validate that the metadata from
each side of the sheet matches with the other. This primarily adds defense in depth against cases where two sheets are stuck together.

Updates the HMPB fixtures to include a ballot for a second precinct in
order to test this logic.