test(scan): stop state machine on test end (#4397)
Before this change, the tests would print warnings that there were resources being held that were preventing NodeJS from exiting within 1s of the tests ending. The state machine's paper status polling was why. This change stops the state machine when the tests are over to prevent that from happening.