feat(scan): decide when to scan on the backend
This is a step in the direction of supporting the PDI scanner in VxScan. The PDI scanner has an async scan mode that means the scanner decides when to scan. To use this mode, we can't have the VxScan frontend deciding when to scan. As a way of moving us closer to that goal, this change moves the decision of when to scan to the backend for the Custom scanner.