VxDesign: Fix bug for primary elections (#4400)
There was a well-intentioned assertion guarding against an edge case in
primary elections where there were no partisan contests (which would
result in no ballot styles being generated). The intention was to
explain to the user why there were no ballot styles in that case.
However, crashing the app until partisan contests are added prevents the
user from being able to address the issue, so it was not a good
solution.