VxAdmin: Improve configure UI and configure on the backend (#4433)
* readElectionPackageFromFile: take a path and return an error

Prepping this to be used on the backend instead of the frontend.

* Consolidate election package functions in libs/backend

* Rename function to readSignedElectionPackageFromUsb

To differentiate reading the signed election package from reading an
unsigned election package.

* VxAdmin: Improve configure UI and read files on backend

- List potential election package zip files on screen
- Remove the "Load Demo Election" function
- Allow using a file picker to select either an election.json or
  election-package.zip file
- Load and parse the election configuration files on the backend

* Update frontend tests

* Update backend tests

* Convert libs/utils to absolute coverage thresholds

* Update integration tests

* Fix VxDesign tests

I have no idea why these broke. Maybe some dependency got upgraded. Just
found a quick fix by googling.

* Test coverage

* grout: Identify Dates in more resilient way

For some reason, this was breaking in tests in apps/admin/backend when
trying to serialize file ctime. Probably mismatched JS environments or
something. This is a quick fix.

* Add check for defense-in-depth

* Fix integration tests

---------

Co-authored-by: Arsalan Sufi <arsalan@voting.works>