refactor(scan): simplify `isReadyToScan`
This was originally copied from the frontend to granularly choose which screens to render based on the current state, then returned specific statuses on the backend. Now that we're just returning `true` or `false` and do not have immediate plans for this to drive the frontend we can simplify it a fair amount.