refactor: use `IteratorPlus` in a few more places (#4309)
I looked for places where we were building an array using `.push` inside a `for` loop and found some that I thought could be improved using `IteratorPlus`, either with `iter` directly or one of the helpers like `integers`. There are more, but this seemed like a good starting point of the ones that I felt were fairly clearly better this way.