[VxScan] Decide when to scan on the backend (#4403)
* refactor(scan): extract `constructAuthMachineState`

Also makes it take `Store` instead of `Workspace` since it only ever accesses `workspace.store`.

* test(scan): simulate card removal after opening polls

* feat(scan): decide when to scan on the backend

This is a step in the direction of supporting the PDI scanner in VxScan. The PDI scanner has an async scan mode that means the scanner decides when to scan. To use this mode, we can't have the VxScan frontend deciding when to scan. As a way of moving us closer to that goal, this change moves the decision of when to scan to the backend for the Custom scanner.

* test(scan): use `electionGeneral` fixture for simpler app flow tests

* test(scan): add tests for more "don't scan yet" conditions

* refactor(scan): simplify the ready to scan check

I had envisioned using the app flow state as a way of driving the frontend, but that would take a lot more changes than I should be making in this PR. So I'm turning the check into a simple yes/no.

* refactor(scan): use state machine

Rather than hiding the conditional inside an `invoke` handler we make it part of the state machine.

* refactor(scan): clarify state as `hardware_ready_to_scan`

Since the app may not be ready to scan even if the hardware is, we rename the `ready_to_scan` state as `hardware_ready_to_scan`.

* refactor(scan): simplify `isReadyToScan`

This was originally copied from the frontend to granularly choose which screens to render based on the current state, then returned specific statuses on the backend. Now that we're just returning `true` or `false` and do not have immediate plans for this to drive the frontend we can simplify it a fair amount.