Attach grid position to bubble element in ballot document
This allows us to create test deck ballots in two phases: laying out the
blank ballot and subsequently marking. When laying out the ballot, we
tag the bubbles with their grid position so that we can later detect
them and mark them.

I chose to do it this way (rather than our previous approach of passing
votes to mark while laying out the ballot) as an experiment because I'm
interested in decoupling various parts of the ballot rendering process
more. Now that this data is part of the Document object, we can also
refactor to extract the grid layouts from the Document after ballot
layout (rather than collecting and returning them as a separate value
during ballot layout). I'm hoping this will make future versions of the
ballot layout engine simpler.