VxScan: print from backend rather than frontend (#4506)
* remove the QR code page from precinct tally reports

this added complexity and depended on the no longer working / to-be removed `kiosk.sign` method. @benadida gave the go ahead to remove this.

* update kiosk-browser type

* remove PRECINCT_REPORT_DESTINATION

we will still support printing to an HP printer with VxScan, for development purposes, but we don't need to render different text or print different numbers of copies because we're not going that direction with the product. confirmed with @mattrow

* backend: ability to print tally report and get printer status

* frontend: simplify tally report flow by just calling backend