Attempt to fix some flaky apps/scan/backend tests (#4528)
* Increase delay after accepting before resetting to no_paper

I saw some failures where the state machine moved on to no_paper before
the test could check for the accepted state.

* Wait for scanning status after initiating scan

Since the move of the decision to start scanning to the backend, it's
possible for a momentary pause in the hardware_ready_to_scan state before
transitioning to the scanning state. Using waitForStatus rather than
expectStatus can smooth out that blip.